Error handling

+++ backend/src/services/client.service.ts
@@ -1,17 +1,25 @@
 import type { PrismaClient, Client } from '../generated/prisma/index.js';
+import type { HttpErrors } from '@fastify/sensible';
 import type {
   CreateClientData,
   UpdateClientData,
 } from '../types/client.types.ts';
 
 export class ClientService {
+  private error: HttpErrors;
   private prisma: PrismaClient;
 
-  constructor(prisma: PrismaClient) {
+  constructor(error: HttpErrors, prisma: PrismaClient) {
+    this.error = error;
     this.prisma = prisma;
   }
 
   create = async (data: CreateClientData): Promise<Client> => {
+    const clientFound = await this.prisma.client.findUnique({
+      where: { email: data.email },
+    });
+    if (clientFound) throw this.error.badRequest('Email already exists');
+
     return this.prisma.client.create({ data });
   };
 
@@ -19,8 +27,8 @@ export class ClientService {
     return this.prisma.client.findMany();
   };
 
-  getById = async (id: string): Promise<Client | null> => {
-    return this.prisma.client.findUnique({
+  getById = async (id: string): Promise<Client> => {
+    const clientFound = await this.prisma.client.findUnique({
       where: { id },
       include: {
         goals: true,
@@ -29,9 +37,22 @@ export class ClientService {
         simulations: true,
       },
     });
+    if (!clientFound) throw this.error.notFound('Client not found');
+
+    return clientFound;
   };
 
   update = async (id: string, data: UpdateClientData): Promise<Client> => {
+    const clientFound = await this.prisma.client.findUnique({ where: { id } });
+    if (!clientFound) throw this.error.notFound('Client not found');
+
+    if (data.email !== clientFound.email) {
+      const foundByEmail = await this.prisma.client.findUnique({
+        where: { email: data.email },
+      });
+      if (foundByEmail) throw this.error.badRequest('Email already exists');
+    }
+
     return this.prisma.client.update({
       where: { id },
       data,
@@ -39,6 +60,9 @@ export class ClientService {
   };
 
   delete = async (id: string): Promise<Client> => {
+    const clientFound = await this.prisma.client.findUnique({ where: { id } });
+    if (!clientFound) throw this.error.notFound('Client not found');
+
     return this.prisma.client.delete({
       where: { id },
     });
